Added the database statistics to the front page. Could use some advice as to wear to put it so that it fits nicely. The default Image Map size is now 30 x 50 mm. The default zoom is 1mm per box which fits the Image Map. Fixed zooming mistakes caused by rounding error and similar things. Saving the grid correctly saves the locations of the images. Navigating away from the image map page saves the image map first.

+++ mpdb-client/src/edu/rpi/metpetdb/client/ui/image/browser/PanHandler.java
@@ -49,8 +49,8 @@ public class PanHandler {
 		imageBrowser.updateBoundary();
 		while (itr.hasNext()) {
 			final ImageOnGridContainer iog = itr.next();
-			final int newX = iog.getCurrentContainerPosition().x + deltaX;
-			final int newY = iog.getCurrentContainerPosition().y + deltaY;
+			final double newX = iog.getCurrentContainerPosition().x + deltaX;
+			final double newY = iog.getCurrentContainerPosition().y + deltaY;
 			this.grid.setWidgetPosition(iog.getImageContainer(), newX, newY);
 			iog.pan(deltaX, deltaY);
 		}
